auth part 1

+++ my-app-laravel/routes/web.php
@@ -7,18 +7,41 @@ use App\Http\Controllers\{
     ViaCepController
 };
 
-Route::get('/posts', [PostController::class, 'index'])->name('posts.index');
-Route::get('/users/{id}/posts', [PostController::class, 'show'])->name('posts.show');
+require __DIR__.'/auth.php';
 
-Route::get('/users', [UserController::class, 'index'])->name('users.index');
-Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
-Route::post('/user', [UserController::class, 'store'])->name('users.store');
-Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit');
-Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update');
-Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy');
-Route::get('/users/{id}', [UserController::class, 'show'])->name('users.show');
+Route::get('/', function () {
+    return view('home');
+});
 
-//VIA CEP WEBSERVICE
-Route::get('/viacep', [ViaCepController::class, 'index'])->name('viacep.index');
-Route::post('/viacep', [ViaCepController::class, 'index'])->name('viacep.index.post');
-Route::get('/viacep/{cep}', [ViaCepController::class, 'show'])->name('viacep.show');
+Route::middleware(['auth'])->group(function (){
+    Route::get('/posts', [PostController::class, 'index'])->name('posts.index');
+    Route::get('/users/{id}/posts', [PostController::class, 'show'])->name('posts.show');
+
+    Route::get('/users', [UserController::class, 'index'])->name('users.index')->middleware('auth');
+    Route::get('/users/create', [UserController::class, 'create'])->name('users.create')->middleware('auth');
+    Route::post('/user', [UserController::class, 'store'])->name('users.store');
+    Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit')->middleware('auth');
+    Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update')->middleware('auth');
+    Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy')->middleware('auth');
+    Route::get('/users/{id}', [UserController::class, 'show'])->name('users.show');
+});
+
+// Route::get('/posts', [PostController::class, 'index'])->name('posts.index');
+// Route::get('/users/{id}/posts', [PostController::class, 'show'])->name('posts.show');
+
+// Route::get('/users', [UserController::class, 'index'])->name('users.index');
+// Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
+// Route::post('/user', [UserController::class, 'store'])->name('users.store');
+// Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit');
+// Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update');
+// Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy');
+// Route::get('/users/{id}', [UserController::class, 'show'])->name('users.show');
+
+// //VIA CEP WEBSERVICE
+// Route::get('/viacep', [ViaCepController::class, 'index'])->name('viacep.index');
+// Route::post('/viacep', [ViaCepController::class, 'index'])->name('viacep.index.post');
+// Route::get('/viacep/{cep}', [ViaCepController::class, 'show'])->name('viacep.show');
+
+// Route::get('/dashboard', function () {
+//     return view('dashboard');
+// })->middleware(['auth'])->name('dashboard');
